+++ lib/mobx.dart
@@ -1,0 +1,8 @@
+library sliver_glue.mobx;
+
+import "package:flutter/material.dart";
+import "package:mobx/mobx.dart";
+
+import "sliver_glue.dart";
+
+part "src/mobx/sliver_glue_observable_list.dart";